app/src/requirements.txt
# Flask and web framework dependencies
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Database drivers
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.7

# Metrics and monitoring
prometheus-flask-exporter==0.23.0
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# HTTP client
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Testing (optional, for development)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code quality (optional, for development)
flake8==6.1.0
black==23.12.1

# Production server
gunicorn==21.2.0
gevent==23.9.1